Fix scanner message storm caused by #916
